fix: resolve PostgreSQL type inference error in upsert_domain_metadata
The SQL used $5 in both integer context (successful_scrapes + $5) and
float context ($5::float / $4), causing asyncpg's prepared statement
to fail with "inconsistent types deduced for parameter $5: double
precision versus integer". Every URL processed by ContentWorker hit
this error, resulting in zero data extraction.

Fix: explicitly cast $4 and $5 as ::int in all integer contexts so
PostgreSQL can unambiguously resolve parameter types.

Also add default=str to json.dumps in batch_insert_scraped_data as a
safety net for non-serializable metadata values.

+++ src/db/queries/domains.py
@@ -27,18 +27,18 @@ async def upsert_domain_metadata(
             domain, last_successful_strategy, block_count,
             last_scraped_at, total_scrapes, successful_scrapes, success_rate
         )
-        VALUES ($1, $2, $3, NOW(), $4, $5,
+        VALUES ($1, $2, $3, NOW(), $4::int, $5::int,
                 CASE WHEN $4 > 0 THEN $5::float / $4 ELSE NULL END)
         ON CONFLICT (domain) DO UPDATE SET
             last_successful_strategy = COALESCE($2, domain_metadata.last_successful_strategy),
             block_count = domain_metadata.block_count + $3,
             last_scraped_at = NOW(),
-            total_scrapes = domain_metadata.total_scrapes + $4,
-            successful_scrapes = domain_metadata.successful_scrapes + $5,
+            total_scrapes = domain_metadata.total_scrapes + $4::int,
+            successful_scrapes = domain_metadata.successful_scrapes + $5::int,
             success_rate = CASE
-                WHEN (domain_metadata.total_scrapes + $4) > 0
-                THEN (domain_metadata.successful_scrapes + $5)::float
-                     / (domain_metadata.total_scrapes + $4)
+                WHEN (domain_metadata.total_scrapes + $4::int) > 0
+                THEN (domain_metadata.successful_scrapes + $5::int)::float
+                     / (domain_metadata.total_scrapes + $4::int)
                 ELSE domain_metadata.success_rate
             END,
             updated_at = NOW()
